refactor: split ether function
