+++ Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 
 [lib]
 path = "src/lib.rs"
+name = "sign_tx_lib"
 crate-type = ["cdylib"]
 
 [dependencies]
